Added initial TrackPoint softbutton support
Many of the new ThinkPad laptops made by Lenovo feature a trackpoint
that instead of having it's own physical click buttons, rely on a
secondary area of softbuttons on the touchpad itself. Lenovo's intended
behavior for this is for the trackpad to stop tracking the motion on the
clickpad and make use of the secondary softbutton area whenever the
trackpoint is in use. These changes add the capabilities required by
the driver to make this possible.